Update to compression (#12)
* Reworked compression so that each class implements an interface

* Consolidating setup functions

* Small typo
